pyproject.toml
[project]
name = "nemesis-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "fastapi>=0.115.12",
    "groq>=0.4.1",
    "isort>=6.0.1",
    "jsonpatch>=1.33",
    "openai>=1.82.0",
    "pocketflow>=0.0.2",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.31.0",
    "ruff>=0.11.11",
    "uvicorn>=0.34.2",
]
